app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Kamran Abdi <kamranabdicse@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
gunicorn = "^20.1.0"
pydantic = {extras = ["dotenv", "email"], version = "^1.10.4"}
psycopg2-binary = "^2.9.5"
alembic = "^1.9.1"
SQLAlchemy = "^1.4.45"
celery = "^5.2.7"
fastapi = "^0.89.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
persiantools = "^3.0.1"
python-multipart = "^0.0.5"
tenacity = "^8.1.0"
rocketry = "^2.5.1"
fakeredis = "^2.10.0"
aioredis = "^2.0.1"
redis = "^4.5.1"
asyncpg = "^0.27.0"
flake8 = "^6.0.0"
black = "^23.3.0"
httpx = "^0.23.3"
jdatetime = "^4.1.1"
pandas = "^2.0.0"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
